Ejercicio boletin match

+++ Phyton/Ejercicios casa/Vacaciones.py
@@ -1,4 +1,4 @@
-AñosTrabajados = int(input("Cuantos años llevas trabajados? "))
+AñosTrabajados = float(input("Cuantos años llevas trabajados? "))
 
 if AñosTrabajados > 0 :
     print ("Tienes 30 dias de vacaciones")
